+++ src/domain/restaurants.ts
@@ -1,4 +1,4 @@
-import { CATEGORY_ALL } from "./../constants/restaurants";
+import { CATEGORY_ALL } from "../constants/restaurants";
 import { BY_DISTANCE, BY_NAME } from "../constants/restaurants";
 import { AlignFilter, CategoryFilter, Restaurant } from "../types/restaurants";
